chore: revert dependabot clone
