--- .github/workflows/update-actions.yml
@@ -1,21 +1,0 @@
-name: Update pinned GitHub Action versions
-
-on:
-  schedule:
-    - cron: '0 7 * * 1,3' # Every Monday and Wednesday at 07:00 UTC
-  workflow_dispatch:
-
-jobs:
-  update:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-      pull-requests: write
-
-    steps:
-      - uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
-
-      - uses: shopware/github-actions/update-pinned-actions@main
-        with:
-          github-token: ${{ github.token }}
-          reviewers: shopware/product-operations
